Dev3 (SPRINT 3) (#5)
* instalar typeOrm

* Instalar y configurar typeORM

* Se implemento getById en carton.controller y se implemento tabla casilleros

* Se creo la tabla filas con sus funcionalidades

* Puesta en funsión de la tabla imagenes con sus métodos

* Se corrigieron los datos de prueba de los cartones

* Refactorización: se marcó métodos como asincrónicos y se definió tipos de retorno

* Creación Métodos POST y getByID

* Metodos SPRINT 2

* Subida rama Juanma Sprint 2

* Cambios en casillero.controllers, casillero.service, fila.controller, imagen.controller, y imagen.service

* Implementacion de Delete-Patch y Response en Usuario/Logueo/Registro/Desempeno

* DELETE, PUT y Response

* Se actualizo controladores y servicios

* Cambios para el nuevo repo

* Codigos de respuesta sobre premios y puntajes

* Agregados Métodos Delete y Patch + Response codes

* Metodos Delete-Patch y codigos Response en stockPremios

* Correciones al Sprint 3

* Correcciones DEV3

---------

+++ src/metodos_pago/metodospago.service.ts
@@ -10,7 +10,7 @@ const baseUrl = 'http://localhost:3030/metodospago'
 export class MetodosPagoService {
   async create(createMetodosPagoDto: CreateMetodosPagoDto): Promise<MetodosPago> {
     const datos = await this.findAll();
-    const id = datos.length ? setId(datos[datos.length - 1].id) : setId(0);
+    const id = datos.length ? setId(datos[datos.length - 1].id).toString() : setId(0);
     const newMetodoPago = { ...createMetodosPagoDto, id };
     const res = await fetch(baseUrl, {
       method: 'POST',
@@ -31,25 +31,32 @@ export class MetodosPagoService {
 
   async findOne(id: number): Promise<MetodosPago> {
     const res = await fetch(`${baseUrl}/${id}`);
+    if(!res.ok) return;
     const parsed = await res.json();
     return parsed;
   }
 
   async update(id: number, updateMetodosPagoDto: UpdateMetodosPagoDto): Promise<MetodosPago> {
+    const isMetodo = await this.findOne(id);
+    if(!isMetodo)return;
     const res = await fetch(`${baseUrl}/${id}`, {
       method: 'PATCH',
       headers: {
         'Content-Type': 'application/json',
       },
       body: JSON.stringify(updateMetodosPagoDto),
-    });
+    });    
     const parsed = await res.json();
     return parsed;
   }
 
-  async remove(id: number): Promise<void> {
-    await fetch(`${baseUrl}/${id}`, {
+  async remove(id: number): Promise<MetodosPago> {
+    const isMetodo = await this.findOne(id);
+    if(!isMetodo)return;
+    const res = await fetch(`${baseUrl}/${id}`, {
       method: 'DELETE',
     });
+    const parsed = res.json();
+    return parsed;
   }
 }
