+++ src/premios/premios.service.ts
@@ -5,13 +5,13 @@ import { Premios } from './entities/premios.entity';
 import { setId } from 'src/funciones/funciones';
 
 
-const baseUrl = 'http://localhost:3030/premios/'
+const baseUrl = 'http://localhost:3030/premios'
 
 @Injectable()
 export class PremiosService {
   async create(createPremiosDto: CreatePremiosDto): Promise<Premios> {
     const datos = await this.findAll();
-    const id = datos.length ? setId(datos[datos.length - 1].id) : setId(0);
+    const id = datos.length ? setId(datos[datos.length - 1].id).toString() : setId(0);
     const newPremio = { ...createPremiosDto, id };
     const res = await fetch(baseUrl, {
       method: 'POST',
@@ -32,11 +32,14 @@ export class PremiosService {
 
   async findOne(id: number): Promise<Premios> {
     const res = await fetch(`${baseUrl}/${id}`);
+    if(!res.ok)return;
     const parsed = await res.json();
     return parsed;
   }
 
   async update(id: number, updatePremiosDto: UpdatePremiosDto): Promise<Premios> {
+    const isPremio = await this.findOne(id);
+    if(!isPremio)return;
     const res = await fetch(`${baseUrl}/${id}`, {
       method: 'PATCH',
       headers: {
@@ -48,9 +51,13 @@ export class PremiosService {
     return parsed;
   }
 
-  async remove(id: number): Promise<void> {
-    await fetch(`${baseUrl}/${id}`, {
-      method: 'DELETE',
+  async remove(id: number){
+    const isPremio = await this.findOne(id);
+    if(!isPremio)return;
+    const res = await fetch(`${baseUrl}/${id}`,{
+      method: "DELETE",      
     });
-  }
+    const parsed = res.json();
+    return parsed;        
+  };
 }
